Changing XRRMonitorInfo.Outputs from IntPtr to IntPtr*
Change XRRMonitorInfo.Outputs from IntPtr to IntPtr* and iterate over
all Outputs to detect the smallest pixeldensity.

+++ src/Avalonia.X11/XLib.cs
@@ -502,10 +502,10 @@ namespace Avalonia.X11
             XRRGetMonitors(IntPtr dpy, IntPtr window, bool get_active, out int nmonitors);
 
         [DllImport(libX11Randr)]
-        public static extern IntPtr* XRRListOutputProperties(IntPtr dpy, long output, out int count);
+        public static extern IntPtr* XRRListOutputProperties(IntPtr dpy, IntPtr output, out int count);
 
         [DllImport(libX11Randr)]
-        public static extern int XRRGetOutputProperty(IntPtr dpy, long output, IntPtr atom, long offset, long length, bool _delete, bool pending, IntPtr req_type, out int actual_type, out int actual_format, out int nitems, out long bytes_after, out IntPtr prop);
+        public static extern int XRRGetOutputProperty(IntPtr dpy, IntPtr output, IntPtr atom, int offset, int length, bool _delete, bool pending, IntPtr req_type, out IntPtr actual_type, out int actual_format, out int nitems, out long bytes_after, out IntPtr prop);
             
         [DllImport(libX11Randr)]
         public static extern void XRRSelectInput(IntPtr dpy, IntPtr window, RandrEventMask mask);
